requirements.txt
annotated-types==0.7.0

asgiref==3.11.1

certifi==2026.2.25

cffi==2.0.0

charset-normalizer==3.4.5

colorama==0.4.6

crispy-bootstrap5==2026.3

cryptography==46.0.5

dj-database-url==3.1.2

Django==6.0.3

django-allauth>=65.13.1,<66

django-crispy-forms==2.6

PyJWT>=2.8.0

google-ai-generativelanguage==0.6.15

google-api-core==2.25.2

google-api-python-client==2.192.0

google-auth==2.49.1

google-auth-httplib2==0.3.0

google-generativeai==0.8.6

google-genai==1.73.1

googleapis-common-protos==1.73.0

grpcio==1.78.0

grpcio-status==1.71.2

gunicorn==25.1.0

httplib2==0.31.2

idna==3.11

numpy==2.4.3

packaging==26.0

pandas==3.0.1

pillow==12.1.1

proto-plus==1.27.1

protobuf==5.29.6

psycopg2-binary==2.9.11

pyasn1==0.6.2

pyasn1_modules==0.4.2

pycparser==3.0

pydantic==2.12.5

pydantic_core==2.41.5

pyparsing==3.3.2

python-dateutil==2.9.0.post0

python-dotenv==1.2.2

requests==2.32.5

six==1.17.0

sqlparse==0.5.5

stripe==15.0.1

tqdm==4.67.3

typing-inspection==0.4.2

typing_extensions==4.15.0

tzdata==2025.3

uritemplate==4.2.0

urllib3==2.6.3

whitenoise==6.12.0


